move query conversion into conjureFetch
