+++ src/codegen/calculateTemplatedUrlForEndpoint.ts
@@ -1,52 +1,7 @@
-import type { IArgumentDefinition, IEndpointDefinition, IParameterType_Query } from "conjure-api";
+import type { IEndpointDefinition } from "conjure-api";
 
 export function calculateTemplatedUrlForEndpoint(endpoint: IEndpointDefinition) {
   const args = endpoint.args ?? [];
-  const queryArgs = args.filter(isQueryArgument) ?? [];
-  const queryPortion = queryArgs.length === 0
-    ? ""
-    : "?${" + `new URLSearchParams(Object.entries({ ${
-      queryArgs.map(a => {
-        // a.type.type === "primitive"  && a.type.primitive === ""
-        switch (a.type.type) {
-          case "map":
-            throw new Error(
-              `Unsupported type ${a.type.type} while generating ${endpoint.endpointName}`,
-            );
-          case "list":
-          case "set":
-            return `"${a.paramType.query.paramId}": ${a.argName}.join(",")`;
-
-          case "reference":
-            return `"${a.paramType.query.paramId}": ${a.argName}`;
-          case "primitive":
-            switch (a.type.primitive) {
-              case "STRING":
-              case "BEARERTOKEN":
-              case "RID":
-              case "UUID":
-                return `"${a.paramType.query.paramId}": ${a.argName}`;
-            }
-        }
-
-        return `"${a.paramType.query.paramId}": ${a.argName}`;
-      }).join(",")
-      // omit undefined or empty query values
-    } }).reduce(
-      (acc, [key, value]) => {
-        if (value == null) {
-          return acc;
-        }
-        const paramValue = "" + value;
-        if (paramValue.length === 0) {
-          return acc;
-        }
-        acc[key] = paramValue;
-        return acc;
-      },
-      {} as { [key: string]: string },
-    )` + ")}";
-
   return "`" + args.reduce(
     (p, c) => {
       if (c.paramType.type === "path") {
@@ -55,11 +10,5 @@ export function calculateTemplatedUrlForEndpoint(endpoint: IEndpointDefinition) 
       return p;
     },
     endpoint.httpPath,
-  ) + queryPortion + "`";
-}
-
-function isQueryArgument(
-  a: IArgumentDefinition,
-): a is IArgumentDefinition & { paramType: IParameterType_Query } {
-  return a.paramType.type === "query";
+  ) + "`";
 }
